+++ polymarket_pandas/__init__.py
@@ -1,3 +1,8 @@
+from polymarket_pandas.async_client import AsyncPolymarketPandas
+from polymarket_pandas.async_ws import (
+    AsyncPolymarketWebSocket,
+    AsyncPolymarketWebSocketSession,
+)
 from polymarket_pandas.client import PolymarketPandas
 from polymarket_pandas.exceptions import (
     PolymarketAPIError,
@@ -8,6 +13,9 @@ from polymarket_pandas.exceptions import (
 from polymarket_pandas.ws import PolymarketWebSocket, PolymarketWebSocketSession
 
 __all__ = [
+    "AsyncPolymarketPandas",
+    "AsyncPolymarketWebSocket",
+    "AsyncPolymarketWebSocketSession",
     "PolymarketPandas",
     "PolymarketWebSocket",
     "PolymarketWebSocketSession",